feat: implementation of auth middleware and separating funcionalities
